+++ src/search_disk_api.h
@@ -83,7 +83,7 @@ typedef struct DocTableDiskAPI {
    * @param maxFreq Maximum term frequency in the document
    * @return New document ID, or 0 on error/duplicate
    */
-  t_docId (*putDocument)(RedisSearchDiskIndexSpec* handle, const char* key, size_t keyLen, double score, uint32_t flags, uint32_t maxFreq);
+  t_docId (*putDocument)(RedisSearchDiskIndexSpec* handle, const char* key, size_t keyLen, float score, uint32_t flags, uint32_t maxFreq);
 
   /**
    * @brief Returns whether the docId is in the deleted set
